+++ CHANGELOG.md
@@ -95,6 +95,8 @@
   [#1009](https://github.com/opendatateam/udata/issues/1009)
 - Reindex datasets when their parent organization is purged
   [#1008](https://github.com/opendatateam/udata/issues/1008)
+- Fix image placeholders display in various places
+  [#1006](https://github.com/opendatateam/udata/issues/1006)
 
 ## 1.0.11 (2017-05-25)